Added Freezie
With this year's wintersday comes a new raid/dungeon-like instance: Secret Lair of the Snowmen. The only notable part to add in the Raid section is the final (and only) boss: Freezie. The radius starts right after you walk through the door to make sure it encompasses the entire fight. Added this to the bottom of the "raid" section, as well as added the instance's ID to the list towards the top.